Created StringBuilderUtil
- Util class to be used for dealing with building Strings or using
StringBuilder
- Moved String building logic present in StringUtil to StringBuilderUtil
- Updated FileUtil and FormatNode, which used the StringUtil.build
methods, to use them in StringBuilderUtil now
